+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/VenusContext.java
@@ -27,6 +27,9 @@ public class VenusContext {
     @Getter @Setter
     private boolean addDocFacace;
 
+    @Getter @Setter
+    private boolean force;
+
     @Getter
     private List<String> errors;